Move calling the notification handler to the rate limit check
